fix(SUP-44191):v7 player issues when using picture-in-picture
